Refactor admin routes and UI modules
